Conflicts:
	Common/Header/Device/device.h
	Common/Source/Device/devVega.cpp

+++ Common/Source/LoggerSign.cpp
@@ -49,10 +49,7 @@
 #include "Device/device.h"
 #include "InputEvents.h"
 #include "Compatibility/string.h"
-
-#ifdef GNAV
 #include "UtilsSystem.hpp" // for FileExistsW()
-#endif
 
 HINSTANCE GRecordDLLHandle = NULL;
 
@@ -154,10 +151,8 @@ LoggerImpl::IGCWriteRecord(const char *szIn, const TCHAR* szLoggerFileName)
       for (i = 0; (i <= iLen) && (i < MAX_IGC_BUFF); i++)
 	buffer[i] = (TCHAR)charbuffer[i];
 
-#ifndef _SIM_
-      if (LoggerGActive())
+      if (!is_simulator() && LoggerGActive())
 	GRecordAppendRecordToBuffer(pbuffer);
-#endif
 
       FlushFileBuffers(hFile);
       CloseHandle(hFile);
@@ -187,17 +182,18 @@ void LoggerImpl::LinkGRecordDLL(void)
       bFirstTime=false;
 
       StartupStore(TEXT("Searching for GRecordDLL\n"));
-#ifdef GNAV
-      if (FileExistsW(TEXT("\\NOR Flash\\GRecordDLL.dat"))) {
-	StartupStore(TEXT("Updating GRecordDLL.DLL\n"));
-	DeleteFile(TEXT("\\NOR Flash\\GRecordDLL.DLL"));
-	MoveFile(TEXT("\\NOR Flash\\GRecordDLL.dat"),
-		 TEXT("\\NOR Flash\\GRecordDLL.DLL"));
-      }
-      GRecordDLLHandle = LoadLibrary(TEXT("\\NOR Flash\\GRecordDLL.DLL"));
-#else
-      GRecordDLLHandle = LoadLibrary(TEXT("GRecordDLL.DLL"));
-#endif
+      if (is_altair()) {
+        if (FileExistsW(TEXT("\\NOR Flash\\GRecordDLL.dat"))) {
+          StartupStore(TEXT("Updating GRecordDLL.DLL\n"));
+          DeleteFile(TEXT("\\NOR Flash\\GRecordDLL.DLL"));
+          MoveFile(TEXT("\\NOR Flash\\GRecordDLL.dat"),
+                   TEXT("\\NOR Flash\\GRecordDLL.DLL"));
+        }
+
+        GRecordDLLHandle = LoadLibrary(TEXT("\\NOR Flash\\GRecordDLL.DLL"));
+      } else
+        GRecordDLLHandle = LoadLibrary(TEXT("GRecordDLL.DLL"));
+
       if (GRecordDLLHandle != NULL)
         {
 	  BOOL bLoadOK = true;  // if any pointers don't link, disable entire library
@@ -364,9 +360,10 @@ LoggerImpl::LoggerGStop(TCHAR* szLoggerFileName) {
 
 void
 LoggerImpl::LoggerGInit() {
-#ifndef _SIM_
+  if (is_simulator())
+    return;
+
   LinkGRecordDLL(); // try to link DLL if it exists
   if (LoggerGActive())
     GRecordInit();
-#endif
 }
